# api/openapi.yaml
openapi: 3.1.0
info:
  title: Ubiquum Gateway API
  description: High-performance LLM gateway proxy — OpenAI-compatible API with multi-provider routing, virtual keys, spend tracking, and budget enforcement.
  version: 0.1.0
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:4000
    description: Local development
paths:
...
  /v1/responses:
    post:
      summary: Create a model response (OpenAI Responses API)
      description: |
        OpenAI Responses endpoint for Codex CLI/Desktop and the official SDKs.

        Deployments whose provider speaks Responses natively (OpenAI, Azure
        OpenAI, ChatGPT Codex) receive the request unchanged, so the full
        parameter surface — including built-in tools such as `web_search` and
        `code_interpreter` — is available. Every other provider is served
        through a Chat Completions translation; built-in tools are rejected
        with HTTP 400 there rather than silently ignored.
      operationId: createResponse
      tags: [Responses]
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model]
              properties:
                model:
                  type: string
                input:
                  description: A string prompt, or a list of typed input items.
                  oneOf:
                    - type: string
                    - type: array
                      items:
                        type: object
                instructions:
                  type: string
                stream:
                  type: boolean
                background:
                  type: boolean
                  description: Run detached and poll GET /v1/responses/{id}.
                store:
                  type: boolean
                  default: true
                previous_response_id:
                  type: string
                conversation:
                  oneOf:
                    - type: string
                    - type: object
                include:
                  type: array
                  items:
                    type: string
                metadata:
                  type: object
                  additionalProperties:
                    type: string
                reasoning:
                  type: object
                  properties:
                    effort:
                      type: string
                    summary:
                      type: string
                text:
                  type: object
                  description: Structured output configuration (text.format).
                tools:
                  type: array
                  items:
                    type: object
                tool_choice: {}
                parallel_tool_calls:
                  type: boolean
                max_output_tokens:
                  type: integer
                max_tool_calls:
                  type: integer
                temperature:
                  type: number
                top_p:
                  type: number
                top_logprobs:
                  type: integer
                truncation:
                  type: string
                service_tier:
                  type: string
                prompt_cache_key:
                  type: string
                safety_identifier:
                  type: string
                user:
                  type: string
      responses:
        "200":
          description: Response object (non-streaming) or SSE stream (streaming)
          content:
            application/json:
              schema:
                type: object
            text/event-stream:
              schema:
                type: string
                description: SSE stream of typed Responses events
        "400":
          description: Invalid request, or a built-in tool unsupported by the model
        "403":
          description: Model not allowed for this key
        "404":
          description: Model not found
        "429":
          description: Rate limited
  /v1/responses/{response_id}:
    parameters:
      - name: response_id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Retrieve a stored response
      description: |
        Served from the gateway's own store. When the turn was produced by a
        provider with native stateful endpoints, the provider is polled first
        so background responses reflect their current status.

        With `stream=true` the turn is followed as SSE instead of returned as a
        single object, which is how a client resumes a background response whose
        connection dropped: pass the last `sequence_number` it received as
        `starting_after`.
      operationId: getResponse
      tags: [Responses]
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: stream
          in: query
          description: Follow the turn as Server-Sent Events.
          schema:
            type: string
            enum: ["true"]
        - name: starting_after
          in: query
          description: Resume after this sequence number. Requires stream=true.
          schema:
            type: integer
            minimum: 0
        - name: include
          in: query
          description: Forwarded to providers with a native Responses API.
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          description: Response object
          content:
            application/json:
              schema:
                type: object
        "404":
          description: Response not found in this tenant scope
    delete:
      summary: Delete a stored response
      operationId: deleteResponse
      tags: [Responses]
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      responses:
        "200":
          description: Deletion confirmation
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  object:
                    type: string
                  deleted:
                    type: boolean
        "404":
          description: Response not found in this tenant scope
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Virtual API key (sk-ubq-xxx) or master key
    MasterKey:
      type: http
      scheme: bearer
      description: Master key (admin access only)
...